meta - transactions - create a transactions dir in controller and move relevant files into it
