+++ app/scripts/controllers/transactions/lib/tx-state-history-helper.js
