+++ app/scripts/controllers/transactions/nonce-tracker.js
