+++ app/scripts/controllers/transactions/tx-gas-utils.js
@@ -3,7 +3,7 @@ const {
   hexToBn,
   BnMultiplyByFraction,
   bnToHex,
-} = require('./util')
+} = require('../../lib/util')
 const { addHexPrefix } = require('ethereumjs-util')
 const SIMPLE_GAS_COST = '0x5208' // Hex for 21000, cost of a simple send.
