+++ test/unit/nonce-tracker-test.js
@@ -1,5 +1,5 @@
 const assert = require('assert')
-const NonceTracker = require('../../app/scripts/lib/nonce-tracker')
+const NonceTracker = require('../../app/scripts/controllers/transactions/nonce-tracker')
 const MockTxGen = require('../lib/mock-tx-gen')
 let providerResultStub = {}
