+++ test/unit/tx-controller-test.js
@@ -5,7 +5,7 @@ const EthjsQuery = require('ethjs-query')
 const ObservableStore = require('obs-store')
 const sinon = require('sinon')
 const TransactionController = require('../../app/scripts/controllers/transactions')
-const TxGasUtils = require('../../app/scripts/lib/tx-gas-utils')
+const TxGasUtils = require('../../app/scripts/controllers/transactions/tx-gas-utils')
 const { createTestProviderTools } = require('../stub/provider')
 
 const noop = () => true
@@ -208,99 +208,6 @@ describe('Transaction Controller', function () {
       })
       .catch(done)
     })
-  })
-
-  describe('#_validateTxParams', function () {
-    it('does not throw for positive values', function () {
-      var sample = {
-        from: '0x1678a085c290ebd122dc42cba69373b5953b831d',
-        value: '0x01',
-      }
-      txController._validateTxParams(sample)
-    })
-
-    it('returns error for negative values', function () {
-      var sample = {
-        from: '0x1678a085c290ebd122dc42cba69373b5953b831d',
-        value: '-0x01',
-      }
-      try {
-        txController._validateTxParams(sample)
-      } catch (err) {
-        assert.ok(err, 'error')
-      }
-    })
-  })
-
-  describe('#_normalizeTxParams', () => {
-    it('should normalize txParams', () => {
-      let txParams = {
-        chainId: '0x1',
-        from: 'a7df1beDBF813f57096dF77FCd515f0B3900e402',
-        to: null,
-        data: '68656c6c6f20776f726c64',
-        random: 'hello world',
-      }
-
-      let normalizedTxParams = txController._normalizeTxParams(txParams)
-
-      assert(!normalizedTxParams.chainId, 'their should be no chainId')
-      assert(!normalizedTxParams.to, 'their should be no to address if null')
-      assert.equal(normalizedTxParams.from.slice(0, 2), '0x', 'from should be hexPrefixd')
-      assert.equal(normalizedTxParams.data.slice(0, 2), '0x', 'data should be hexPrefixd')
-      assert(!('random' in normalizedTxParams), 'their should be no random key in normalizedTxParams')
-
-      txParams.to = 'a7df1beDBF813f57096dF77FCd515f0B3900e402'
-      normalizedTxParams = txController._normalizeTxParams(txParams)
-      assert.equal(normalizedTxParams.to.slice(0, 2), '0x', 'to should be hexPrefixd')
-
-    })
-  })
-
-  describe('#_validateRecipient', () => {
-    it('removes recipient for txParams with 0x when contract data is provided', function () {
-      const zeroRecipientandDataTxParams = {
-        from: '0x1678a085c290ebd122dc42cba69373b5953b831d',
-        to: '0x',
-        data: 'bytecode',
-      }
-      const sanitizedTxParams = txController._validateRecipient(zeroRecipientandDataTxParams)
-      assert.deepEqual(sanitizedTxParams, { from: '0x1678a085c290ebd122dc42cba69373b5953b831d', data: 'bytecode' }, 'no recipient with 0x')
-    })
-
-    it('should error when recipient is 0x', function () {
-      const zeroRecipientTxParams = {
-        from: '0x1678a085c290ebd122dc42cba69373b5953b831d',
-        to: '0x',
-      }
-      assert.throws(() => { txController._validateRecipient(zeroRecipientTxParams) }, Error, 'Invalid recipient address')
-    })
-  })
-
-
-  describe('#_validateFrom', () => {
-    it('should error when from is not a hex string', function () {
-
-      // where from is undefined
-      const txParams = {}
-      assert.throws(() => { txController._validateFrom(txParams) }, Error, `Invalid from address ${txParams.from} not a string`)
-
-      // where from is array
-      txParams.from = []
-      assert.throws(() => { txController._validateFrom(txParams) }, Error, `Invalid from address ${txParams.from} not a string`)
-
-      // where from is a object
-      txParams.from = {}
-      assert.throws(() => { txController._validateFrom(txParams) }, Error, `Invalid from address ${txParams.from} not a string`)
-
-      // where from is a invalid address
-      txParams.from = 'im going to fail'
-      assert.throws(() => { txController._validateFrom(txParams) }, Error, `Invalid from address`)
-
-      // should run
-      txParams.from ='0x1678a085c290ebd122dc42cba69373b5953b831d'
-      txController._validateFrom(txParams)
-      })
   })
 
   describe('#addTx', function () {
