+++ test/unit/tx-state-history-helper-test.js
@@ -1,6 +1,6 @@
 const assert = require('assert')
 const clone = require('clone')
-const txStateHistoryHelper = require('../../app/scripts/lib/tx-state-history-helper')
+const txStateHistoryHelper = require('../../app/scripts/controllers/transactions/lib/tx-state-history-helper')
 
 describe('deepCloneFromTxMeta', function () {
   it('should clone deep', function () {
